[feat/#29]:popup페이지 생성
