+++ src/app/ui/input/page.jsx
@@ -1,0 +1,23 @@
+"use client";
+import InputTextBox from "@/src/ui/Components/atoms/InputTextBox";
+import { useState } from "react";
+
+const Page = () => {
+  const [placeholderValue, setPlaceholderValue] = useState("값을 입력하세요");
+  const handlePlaceholderChange = (newPlaceholder) => {
+    setPlaceholderValue(newPlaceholder);
+  };
+  return (
+    <>
+      <div className="p-6">
+        <h1 className="text-2xl mb-4">Example Usage of CommonComponent</h1>
+        <InputTextBox
+          height={60}
+          placeholder={placeholderValue}
+          onPlaceholderChange={handlePlaceholderChange}
+        />
+      </div>
+    </>
+  );
+};
+export default Page;
